# specs/001-assessment-system-baseline/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Assessly Backend API
  description: |
    REST API for Assessly flexible assessment system baseline.
    
    **Authentication**:
    - JWT tokens for creators and reviewers (Authorization: Bearer <token>)
    - Secure tokens for participant test access (URL parameter or header)
    
    **Roles**:
    - Creator: Can create/manage own tests
    - Reviewer: Can view all published tests and add manual scores
    - Participant: Anonymous users taking tests (no authentication)
  version: 1.0.0
  contact:
    name: Assessly Team
  license:
    name: Private
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.assessly.example.com/api/v1
    description: Production (placeholder)
...
security:
  - BearerAuth: []
paths:
...
  /tests:
    get:
      tags: [Tests]
      summary: List tests
      description: |
        List tests based on user role:
        - Creators: Only their own tests
        - Reviewers: All published tests
      operationId: listTests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, all]
            default: all
          description: Filter by publish status (creators only)
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Tests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tests:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Tests]
      summary: Create new test
      description: Create a new test (draft) with title and description. Add questions separately.
      operationId: createTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: Essay Writing 101
                description:
                  type: string
                  maxLength: 5000
                  example: Write thoughtful essays on programming concepts
                allow_retakes:
                  type: boolean
                  default: false
                  description: Whether participants can retake this test
      responses:
        '201':
          description: Test created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only creators can create tests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /tests/{testId}/publish:
    post:
      tags: [Tests]
      summary: Publish test
      description: Publish test to make it available to participants (requires at least 1 question)
      operationId: publishTest
      parameters:
        - $ref: '#/components/parameters/TestId'
      responses:
        '200':
          description: Test published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          description: Cannot publish (e.g., no questions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /tests/{testId}/submissions:
    get:
      tags: [Submissions]
      summary: List test submissions
      description: |
        List all submissions for a test.
        - Creators: Only for their own tests
        - Reviewers: For all published tests
      operationId: listSubmissions
      parameters:
        - $ref: '#/components/parameters/TestId'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Submissions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  submissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SubmissionSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================================
  # REVIEW ENDPOINTS
  # ============================================================================
  /reviews/{answerId}:
    put:
      tags: [Reviews]
      summary: Add/update manual review
      description: Reviewer adds or updates manual score and feedback for an answer (FR-020)
      operationId: reviewAnswer
      parameters:
        - name: answerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Answer ID to review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manual_score]
              properties:
                manual_score:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 100
                  example: 85.5
                manual_feedback:
                  type: string
                  maxLength: 10000
                  example: Excellent work! Minor grammar issues on paragraph 3.
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only reviewers can add manual reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Reviews]
      summary: Get review for answer
      description: Get AI and manual review for a specific answer
      operationId: getReview
      parameters:
        - name: answerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Answer or review not found (review may not exist yet)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

# ==============================================================================
# COMPONENTS
# ==============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated users (creators/reviewers)
    
    AccessToken:
      type: apiKey
      in: header
      name: X-Access-Token
      description: Participant access token from email (for submissions)
  parameters:
    TestId:
      name: testId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Test UUID
...
  schemas:
...
    Test:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
        title:
          type: string
          example: Essay Writing 101
        description:
          type: string
          example: Write thoughtful essays on programming concepts
        allow_retakes:
          type: boolean
          example: false
        is_published:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TestSummary:
      type: object
      description: Lightweight test representation for list views
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        is_published:
          type: boolean
        question_count:
          type: integer
          description: Number of questions in test
        submission_count:
          type: integer
          description: Number of submissions received
        created_at:
          type: string
          format: date-time
...
    SubmissionSummary:
      type: object
      description: Lightweight submission representation for list views
      properties:
        id:
          type: string
          format: uuid
        access_email:
          type: string
          format: email
        submitted_at:
          type: string
          format: date-time
        ai_total_score:
          type: number
          format: float
          nullable: true
        manual_total_score:
          type: number
          format: float
          nullable: true
        answer_count:
          type: integer
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        answer_id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
          nullable: true
          description: Reviewer who added manual score (null if AI-only)
        ai_score:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 100
          example: 85.5
        ai_feedback:
          type: string
          nullable: true
          example: Good explanation of separation of concerns. Could elaborate more on dependency inversion.
        ai_scored_at:
          type: string
          format: date-time
          nullable: true
        manual_score:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 100
          example: 90.0
        manual_feedback:
          type: string
          nullable: true
          example: Excellent work! Minor grammar issues on paragraph 3.
        manual_scored_at:
          type: string
          format: date-time
          nullable: true

    # ================================================================
    # UTILITY SCHEMAS
    # ================================================================
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 156
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Invalid email format
        code:
          type: string
          description: Error code for client handling
          example: VALIDATION_ERROR
        details:
          type: object
          description: Additional error context
          additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              email: Invalid email format
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid or expired token
            code: UNAUTHORIZED
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: You do not have permission to access this resource
            code: FORBIDDEN
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Test not found
            code: NOT_FOUND
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: An unexpected error occurred
            code: INTERNAL_ERROR